Move leaderboard delay check
